Allow retargeting normal executions to manual ones

+++ manual_execution.py
@@ -48,6 +48,19 @@ def api_create_manual_scheduler_execution(source_execution_id: str, alias: Optio
     return data["executionId"], data["inputFileUrl"]
 
 
+def api_retarget_manual_scheduler_execution(source_execution_id: str) -> Tuple[str, str]:
+    url = get_api_url("retarget-manual-scheduler-execution")
+    body = {
+        'sourceExecutionId': source_execution_id,
+    }
+
+    response = requests.post(url, headers={'X-API-Key': get_api_key()}, json=body)
+    data = response.json()
+    if response.status_code // 200 > 2:
+        raise RuntimeError(f"API Error: {data['message']}")
+    return data["executionId"], data["inputFileUrl"]
+
+
 def api_generate_manual_scheduler_execution_upload(execution_id: str,
                                                    additional_files: Optional[List[str]] = None) -> Tuple[
     Tuple[str, Dict],
@@ -96,6 +109,26 @@ def create_manual_execution(source_execution_id: str, alias: Optional[str] = Non
         raise RuntimeError("There is an invocation here already")
 
     execution_id, input_file_url = api_create_manual_scheduler_execution(source_execution_id, alias)
+
+    input_data = requests.get(input_file_url).text
+    with open(invocation_dir / 'input.json', 'w') as f:
+        f.write(input_data)
+
+    with open(execution_id_file, 'w') as f:
+        f.write(execution_id)
+
+    print(f"Manual Scheduler Execution {execution_id} is now ready and waiting to receive the processed files")
+
+
+def retarget_manual_execution(source_execution_id: str):
+    invocation_dir = Path().absolute()
+
+    execution_id_file = invocation_dir / '.asp_execution_id'
+    if execution_id_file.exists():
+        raise RuntimeError("There is an invocation here already")
+
+    print("Retargeting and generating input.json; this may take a while...")
+    execution_id, input_file_url = api_retarget_manual_scheduler_execution(source_execution_id)
 
     input_data = requests.get(input_file_url).text
     with open(invocation_dir / 'input.json', 'w') as f:
@@ -163,6 +196,12 @@ def main():
                               help='Alias for this Scheduler Execution',
                               type=str)
 
+    parser_retarget = sub_parsers.add_parser('retarget', help='Retarget a normal Scheduler Execution to a Manual one '
+                                                              'and retrieve input.json')
+    parser_retarget.add_argument('-s', '--source-execution-id',
+                                 help='Source Scheduler Execution ID to retarget and retrieve the input files',
+                                 type=str, required=True)
+
     parser_upload = sub_parsers.add_parser('upload',
                                            help='Upload the generated files (output.json if present and asp.txt ones)')
 
@@ -172,6 +211,8 @@ def main():
     args = parser.parse_args()
     if args.mode == "start":
         create_manual_execution(args.source_execution_id, args.alias)
+    if args.mode == "retarget":
+        retarget_manual_execution(args.source_execution_id)
     elif args.mode == "upload":
         upload_manual_execution()
     elif args.mode == "finish":
